+++ miracle/crypto.py
@@ -3,9 +3,9 @@ from base64 import (
 )
 
 from cryptography.hazmat.backends import default_backend
-from cryptography.hazmat.primitives.asymmetric import padding
-from cryptography.hazmat.primitives import hashes
 from cryptography.hazmat.primitives import serialization
+from jwcrypto.jwe import JWE
+from jwcrypto.jwk import JWK
 
 from miracle import config
 from miracle.log import LOGGER
@@ -22,55 +22,52 @@ class Crypto(object):
 
     _backend = None
     _private = None
+    _private_jwk = None
     _public = None
+    _public_jwk = None
 
     def __init__(self, private_key=None, public_key=None):
         self._backend = default_backend()
         if private_key:
-            self._private = self._load_private(private_key)
+            self._load_private(private_key)
         if public_key:
-            self._public = self._load_public(public_key)
+            self._load_public(public_key)
 
     def _load_private(self, data):
         try:
-            return serialization.load_pem_private_key(
+            self._private = serialization.load_pem_private_key(
                 b64decode(data), password=None, backend=self._backend)
+            self._private_jwk = JWK()
+            self._private_jwk.import_from_pyca(self._private)
         except Exception:  # pragma: no cover
             LOGGER.error(
                 'Failed to parse private key starting with: %s', data[:28])
 
     def _load_public(self, data):
         try:
-            return serialization.load_pem_public_key(
+            self._public = serialization.load_pem_public_key(
                 b64decode(data), backend=self._backend)
+            self._public_jwk = JWK()
+            self._public_jwk.import_from_pyca(self._public)
         except Exception:  # pragma: no cover
             LOGGER.error(
                 'Failed to parse public key starting with: %s', data[:26])
 
     def encrypt(self, plaintext):
         try:
-            ciphertext = self._public.encrypt(
-                plaintext,
-                padding.OAEP(
-                    mgf=padding.MGF1(algorithm=hashes.SHA256()),
-                    algorithm=hashes.SHA256(),
-                    label=None,
-                )
-            )
+            jwe = JWE(plaintext=plaintext,
+                      protected='{"alg":"RSA-OAEP","enc":"A256GCM"}')
+            jwe.add_recipient(self._public_jwk)
+            ciphertext = jwe.serialize(compact=True)
         except Exception:
             raise ValueError("Couldn't encrypt message.")
         return ciphertext
 
     def decrypt(self, ciphertext):
         try:
-            plaintext = self._private.decrypt(
-                ciphertext,
-                padding.OAEP(
-                    mgf=padding.MGF1(algorithm=hashes.SHA256()),
-                    algorithm=hashes.SHA256(),
-                    label=None
-                )
-            )
+            jwe = JWE()
+            jwe.deserialize(ciphertext, key=self._private_jwk)
+            plaintext = jwe.payload
         except Exception:
             raise ValueError("Couldn't decrypt message.")
         return plaintext